Edit functions files. Add new test file.
