+++ src/__tests__/pureFunctions.test.js
@@ -1,0 +1,10 @@
+import {
+  fetchTodoArrayFromLocalStorage,
+  fetchProjectArrayFromLocalStorage,
+  saveTodoArrayInLocalStorage,
+  saveProjectArrayInLocalStorage,
+  deleteTodoObjFromArray,
+  capitalizeFirstLetter,
+  lastId,
+} from '../pureFunctions';
+
